light refactoring (moved Map.context to Transport, whitespaces, deleted  unused functions)
