+++ components/Map/Vehicles/Item/MapVehiclesItem.module.css
@@ -3,8 +3,6 @@
 
     border: none;
     background-color: transparent;
-
-    /* transition: transform 500ms linear; */
 }
 
 .MapVehicleWrapper {
