+++ components/hocs/withMap.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-
 import L from 'leaflet';
 import { useMap } from 'react-leaflet';
