+++ styles/globals.css
@@ -24,6 +24,7 @@ body {
     html {
         color-scheme: dark;
     }
+
     body {
         color: white;
         background: black;
